Utils – Rename parser.py to artifact.py (#33)
Rename src/utils/parser.py → src/utils/artifact.py and
src/utils/tests/test_parser.py → src/utils/tests/test_artifact.py.
Update all imports and documentation references.
