+++ src/utils/__init__.py
@@ -4,6 +4,6 @@
 
 # ** app
 from .lexer import TiferetLexer
-from .parser import ArtifactBlockParser
+from .artifact import ArtifactBlockParser
 from .output import ScanOutputWriter
 from .indent import IndentInjector
